[#166] Preparation for implementing apply of desired state for containers update agent
Some preparation work before the real implementation:
- move parse functions from cli main package to common containerm/util package, to be used also by container update agent implementation.
- added additional validation for container - cannot have a container withcannot have a privileged container with specified devices
